@@ -11,16 +11,16 @@ android {
             keyPassword 'caboLfazl$235711#zH'
         }
     }
-    compileSdkVersion 26
-    buildToolsVersion '28.0.3'
+    compileSdkVersion 29
+    buildToolsVersion '29.0.2'
     defaultConfig {
         applicationId "com.concough.android.concough"
         minSdkVersion 16
-        targetSdkVersion 26
+        targetSdkVersion 29
         multiDexEnabled true
         versionCode 11
         versionName "2.0"
-        testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
+        testInstrumentationRunner "androidx.test.runner.AndroidJUnitRunner"
 //        jackOptions {
 //            enabled true
 //        }
