@@ -6,8 +6,8 @@ import net.danlew.android.joda.JodaTimeAndroid
 import android.app.Activity
 import android.content.Context
 import android.os.Bundle
-import android.support.multidex.MultiDex
-import android.support.multidex.MultiDexApplication
+import androidx.multidex.MultiDex
+import androidx.multidex.MultiDexApplication
 
 
 /**
