@@ -2,59 +2,35 @@ package com.concough.android.concough;
 
 import android.content.Context;
 import android.content.Intent;
-import android.graphics.PorterDuff;
 import android.os.AsyncTask;
-import android.os.Build;
 import android.os.Bundle;
-import android.provider.Settings;
-import android.support.v4.content.ContextCompat;
-import android.support.v7.app.AppCompatActivity;
+
+import androidx.core.content.ContextCompat;
+import androidx.appcompat.app.AppCompatActivity;
 import android.text.Editable;
 import android.text.TextWatcher;
-import android.util.Log;
-import android.view.Gravity;
 import android.view.View;
 import android.widget.Button;
 import android.widget.EditText;
-import android.widget.ProgressBar;
 import android.widget.TextView;
 
 import com.concough.android.general.AlertClass;
-import com.concough.android.models.EntranceModel;
-import com.concough.android.models.EntranceModelHandler;
-import com.concough.android.models.EntranceOpenedCountModelHandler;
-import com.concough.android.models.EntranceQuestionStarredModelHandler;
-import com.concough.android.models.PurchasedModel;
-import com.concough.android.models.PurchasedModelHandler;
 import com.concough.android.rest.AuthRestAPIClass;
-import com.concough.android.rest.DeviceRestAPIClass;
-import com.concough.android.rest.MediaRestAPIClass;
 import com.concough.android.rest.ProfileRestAPIClass;
-import com.concough.android.rest.PurchasedRestAPIClass;
-import com.concough.android.singletons.DeviceInformationSingleton;
 import com.concough.android.singletons.FontCacheSingleton;
 import com.concough.android.singletons.FormatterSingleton;
-import com.concough.android.singletons.MediaCacheSingleton;
 import com.concough.android.singletons.SynchronizationSingleton;
 import com.concough.android.singletons.TokenHandlerSingleton;
 import com.concough.android.singletons.UserDefaultsSingleton;
-import com.concough.android.structures.EntranceStruct;
 import com.concough.android.structures.HTTPErrorType;
 import com.concough.android.structures.NetworkErrorType;
 import com.concough.android.structures.SignupStruct;
 import com.concough.android.utils.KeyChainAccessProxy;
 import com.concough.android.vendor.progressHUD.KProgressHUD;
-import com.google.gson.JsonArray;
-import com.google.gson.JsonElement;
 import com.google.gson.JsonObject;
-import com.google.gson.JsonParser;
 
-import java.io.File;
-import java.io.FileOutputStream;
-import java.util.ArrayList;
 import java.util.Date;
 
-import io.realm.RealmResults;
 import kotlin.Unit;
 import kotlin.jvm.functions.Function0;
 import kotlin.jvm.functions.Function1;
