@@ -2,9 +2,9 @@ package com.concough.android.concough;
 
 import android.content.Context;
 import android.os.Bundle;
-import android.support.v4.app.Fragment;
-import android.support.v7.widget.LinearLayoutManager;
-import android.support.v7.widget.RecyclerView;
+import androidx.fragment.app.Fragment;
+import androidx.recyclerview.widget.LinearLayoutManager;
+import androidx.recyclerview.widget.RecyclerView;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
@@ -51,9 +51,6 @@ public class SectionFragment extends Fragment {
             recyclerView.setAdapter(new SetsAdapter(getContext(), new ArrayList<String>()));
 
             return rootView;
-
-
-
         }
 
     public class SetsAdapter extends RecyclerView.Adapter<RecyclerView.ViewHolder> {
