@@ -22,11 +22,10 @@ package com.concough.android.general;
         import android.view.ScaleGestureDetector;
         import android.view.View;
         import android.view.animation.AccelerateDecelerateInterpolator;
-        import android.widget.ImageView;
         import android.widget.OverScroller;
         import android.widget.Scroller;
 
-public class TouchImageView extends android.support.v7.widget.AppCompatImageView {
+public class TouchImageView extends androidx.appcompat.widget.AppCompatImageView {
 
     private static final String DEBUG = "DEBUG";
 
